Move mf_dgp to examples folder, replace printing with logging
