+++ emukit/multi_fidelity/models/__init__.py
@@ -4,4 +4,3 @@
 
 from .linear_model import GPyLinearMultiFidelityModel  # noqa: F401
 from .non_linear_multi_fidelity_model import NonLinearMultiFidelityModel  # noqa: F401
-from .multi_fidelity_deep_gp import MultiFidelityDeepGP  # noqa: F401
